+++ src/main/java/io/github/realguyman/totally_lit/api/block/NoParticleTorchBlock.java
@@ -1,13 +1,13 @@
-package io.github.realguyman.totally_lit.block;
+package io.github.realguyman.totally_lit.api.block;
 
 import net.minecraft.block.BlockState;
-import net.minecraft.block.WallTorchBlock;
+import net.minecraft.block.TorchBlock;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.random.Random;
 import net.minecraft.world.World;
 
-public class NoParticleWallTorchBlock extends WallTorchBlock {
-    public NoParticleWallTorchBlock(Settings settings) {
+public class NoParticleTorchBlock extends TorchBlock {
+    public NoParticleTorchBlock(Settings settings) {
         super(settings, null);
     }
