+++ src/main/java/io/github/realguyman/totally_lit/api/block/UnlitTorchBlock.java
@@ -1,9 +1,8 @@
-package io.github.realguyman.totally_lit.block;
+package io.github.realguyman.totally_lit.api.block;
 
 import io.github.realguyman.totally_lit.registry.TagRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
-import net.minecraft.block.WallTorchBlock;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
@@ -11,10 +10,10 @@ import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
-public class UnlitWallTorchBlock extends NoParticleWallTorchBlock {
+public class UnlitTorchBlock extends NoParticleTorchBlock {
     private final Block litBlock;
 
-    public UnlitWallTorchBlock(Settings settings, Block litBlock) {
+    public UnlitTorchBlock(Settings settings, Block litBlock) {
         super(settings);
         this.litBlock = litBlock;
     }
@@ -25,7 +24,7 @@ public class UnlitWallTorchBlock extends NoParticleWallTorchBlock {
 
     @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult result) {
-        if (player.getStackInHand(hand).isIn(TagRegistry.TORCH_IGNITER_ITEMS) && world.setBlockState(pos, litBlock.getDefaultState().with(WallTorchBlock.FACING, state.get(WallTorchBlock.FACING)))) {
+        if (player.getStackInHand(hand).isIn(TagRegistry.TORCH_IGNITER_ITEMS) && world.setBlockState(pos, litBlock.getDefaultState())) {
             // TODO: Add a quiet flame sound.
             return ActionResult.SUCCESS;
         }
