+++ src/main/java/io/github/realguyman/totally_lit/api/block/UnlitWallTorchBlock.java
@@ -1,8 +1,9 @@
-package io.github.realguyman.totally_lit.block;
+package io.github.realguyman.totally_lit.api.block;
 
 import io.github.realguyman.totally_lit.registry.TagRegistry;
 import net.minecraft.block.Block;
 import net.minecraft.block.BlockState;
+import net.minecraft.block.WallTorchBlock;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.util.ActionResult;
 import net.minecraft.util.Hand;
@@ -10,10 +11,10 @@ import net.minecraft.util.hit.BlockHitResult;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.world.World;
 
-public class UnlitTorchBlock extends NoParticleTorchBlock {
+public class UnlitWallTorchBlock extends NoParticleWallTorchBlock {
     private final Block litBlock;
 
-    public UnlitTorchBlock(Settings settings, Block litBlock) {
+    public UnlitWallTorchBlock(Settings settings, Block litBlock) {
         super(settings);
         this.litBlock = litBlock;
     }
@@ -24,7 +25,7 @@ public class UnlitTorchBlock extends NoParticleTorchBlock {
 
     @Override
     public ActionResult onUse(BlockState state, World world, BlockPos pos, PlayerEntity player, Hand hand, BlockHitResult result) {
-        if (player.getStackInHand(hand).isIn(TagRegistry.TORCH_IGNITER_ITEMS) && world.setBlockState(pos, litBlock.getDefaultState())) {
+        if (player.getStackInHand(hand).isIn(TagRegistry.TORCH_IGNITER_ITEMS) && world.setBlockState(pos, litBlock.getDefaultState().with(WallTorchBlock.FACING, state.get(WallTorchBlock.FACING)))) {
             // TODO: Add a quiet flame sound.
             return ActionResult.SUCCESS;
         }
