+++ src/main/java/io/github/realguyman/totally_lit/api/item/UnlitTorchItem.java
@@ -1,4 +1,4 @@
-package io.github.realguyman.totally_lit.item;
+package io.github.realguyman.totally_lit.api.item;
 
 import io.github.realguyman.totally_lit.registry.TagRegistry;
 import net.minecraft.block.Block;
