+++ src/main/java/io/github/realguyman/totally_lit/mixin/CampfireBlockMixin.java
@@ -1,6 +1,6 @@
 package io.github.realguyman.totally_lit.mixin;
 
-import io.github.realguyman.totally_lit.item.LitTorchItem;
+import io.github.realguyman.totally_lit.api.item.LitTorchItem;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.BlockWithEntity;
 import net.minecraft.block.CampfireBlock;
