+++ src/main/java/io/github/realguyman/totally_lit/mixin/FlintAndSteelItemMixin.java
@@ -1,6 +1,6 @@
 package io.github.realguyman.totally_lit.mixin;
 
-import io.github.realguyman.totally_lit.block.*;
+import io.github.realguyman.totally_lit.api.block.*;
 import net.minecraft.block.BlockState;
 import net.minecraft.block.LanternBlock;
 import net.minecraft.block.WallTorchBlock;
