+++ src/main/java/io/github/realguyman/totally_lit/mixin/ItemEntityMixin.java
@@ -1,7 +1,7 @@
 package io.github.realguyman.totally_lit.mixin;
 
-import io.github.realguyman.totally_lit.item.LitLanternItem;
-import io.github.realguyman.totally_lit.item.LitTorchItem;
+import io.github.realguyman.totally_lit.api.item.LitLanternItem;
+import io.github.realguyman.totally_lit.api.item.LitTorchItem;
 import io.github.realguyman.totally_lit.registry.ItemRegistry;
 import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityType;
